src/main/java/cd/go/authorization/gitlab/client/CodeChallengeGenerator.java
/*
 * Copyright 2024 ThoughtWorks, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package cd.go.authorization.gitlab.client;

import java.nio.charset.StandardCharsets;
import java.security.*;
import java.util.*;

public class CodeChallengeGenerator {

    private static final SecureRandom RANDOM = new SecureRandom();

    private static String generateCodeVerifier() {
        byte[] codeVerifier = new byte[32];
        RANDOM.nextBytes(codeVerifier);
        return Base64.getUrlEncoder().withoutPadding().encodeToString(codeVerifier);
    }

    public static List<String> generate() {
        String codeVerifier = generateCodeVerifier();
        byte[] bytes = codeVerifier.getBytes(StandardCharsets.US_ASCII);
        MessageDigest messageDigest = null;
        try {
            messageDigest = MessageDigest.getInstance("SHA-256");
        } catch (NoSuchAlgorithmException e) {
            throw new RuntimeException(e);
        }
        messageDigest.update(bytes, 0, bytes.length);
        byte[] digest = messageDigest.digest();
        return List.of(codeVerifier, Base64.getUrlEncoder().withoutPadding().encodeToString(digest));
    }
}
